\name{aerobatStats-package}
\alias{aerobatStats-package}
\alias{aerobatStats}
\docType{package}
\title{
\packageTitle{aerobatStats}
}
\description{
\packageDescription{aerobatStats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aerobatStats}
\packageIndices{aerobatStats}
These are a collection of functions for making computations from aerobatic
contest data.  They encapsulate methods used for any of various statistical
studies of the data.
}
\author{
\packageAuthor{aerobatStats}

Maintainer: \packageMaintainer{Douglas Lovell}
}
\references{
See many documents at fai.org, iac.org
}
\keyword{ aerobatics }
\seealso{
\code{\link[aerobatDB:aerobatDB-package]{aerobatDB}}
}
\examples{
}
